Move pysam function to tools for now, update function callouts in solar_pysam
